+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # mei-friend-online CHANGELOG.md
 
+### 1.0.12 patch
+* Add PlayWright end-to-end tests for the main functionality of mei-friend
+
+### 1.0.11 patch
+* Extend documentation with a new file `CONTRIBUTING.md` to provide guidelines for contributing to mei-friend
+* Update external documentation pages at https://mei-friend.github.io including updated installation instructions, 
+  information on the code structure, and an updated page on contributing to mei-friend
+* Fix bug when loading facsimile source images from URLs starting with raw.githubusercontent (#93)
+
+### 1.0.10 patch
+* Scroll to facsimile image zones corresponding to selected element in editor or notation
+* Make sure zone element edits are reloaded and redrawn in facsimile panel
+* Drawing new zones in non-fullpage mode corrected for Firefox and other browsers
+* New zones remain selected after insertion
+* Clicking on elements with `@facs` attribute will highlight linked zone also in zone edit mode
+
 ### 1.0.9 patch
 * Facsimile panel GUI elements update local storage 
 * Verovio settings retrieved from local storage